// src/api/generated/model/orderItemResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */
import type { OrderItemResponseAttributes } from "./orderItemResponseAttributes.ts";

export interface OrderItemResponse {
	skuId?: string;
	productId?: string;
	title?: string;
	price?: number;
	amount?: number;
	subtotal?: number;
	photo?: string;
	attributes?: OrderItemResponseAttributes;
}