// src/api/generated/model/orderResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */
import type { AddressResponse } from "./addressResponse.ts";
import type { OrderItemResponse } from "./orderItemResponse.ts";

export interface OrderResponse {
	id?: string;
	clientId?: string;
	clientName?: string;
	address?: AddressResponse;
	totalValue?: number;
	status?: string;
	createdAt?: string;
	canceledAt?: string;
	items?: OrderItemResponse[];
}